Changes Borrow.Fine data type from float to decimal

+++ libdb-dotnet/Domain/Borrow.cs
@@ -19,7 +19,7 @@ namespace libdb_dotnet.Domain
 
         public DateOnly? ReturnDate { get; set; }
 
-        [Range(0, float.MaxValue)]
-        public float Fine { get; set; }
+        [Range(0, (double)decimal.MaxValue)]
+        public decimal Fine { get; set; }
     }
 }
